# openapi.yaml
openapi: 3.1.0
info:
  title: CiteFlowAI Agent Research API
  version: "1.0.0"
  description: >
    Agent-payable research endpoint. Any x402 client, human or autonomous, pays
    in USDC via a gasless signed authorization and receives a grounded, cited
    answer. The agent evaluates only ownership-verified, registered sources;
    every source actually cited in the answer is paid at the moment of use,
    and any unspent portion of the budget is refunded to the paying wallet.
    No login and no API key are required, payment is the credential.
  contact:
    name: CiteFlowAI
    url: https://citeflowai.xyz
  license:
    name: MIT

servers:
  - url: https://citeflowai.xyz
    description: Production (Arc Testnet)

paths:
  /api/agent/research:
    get:
      operationId: runResearch
      summary: Run a grounded, cited research query, paid via x402
      description: >
        Requires a valid x402 payment authorization for $1.00 USDC on Arc
        Testnet (eip155:5042002), submitted via the PAYMENT-SIGNATURE header
        (X-PAYMENT is also accepted). A request without a valid authorization
        receives a 402 Payment Required challenge describing how to pay,
        per the x402 protocol (see https://x402.org).
      parameters:
        - name: q
          in: query
          required: true
          description: The research question. Minimum 5 characters.
          schema:
            type: string
            minLength: 5
        - name: PAYMENT-SIGNATURE
          in: header
          required: false
          description: >
            Signed x402 payment authorization (EIP-3009 exact-scheme, Circle
            Gateway batched settlement). Omit to receive the 402 challenge
            describing payment requirements. X-PAYMENT is accepted as an
            alias for this header.
          schema:
            type: string
      responses:
        '200':
          description: Grounded answer, with any sources actually cited paid in USDC.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchResult'
              examples:
                success:
                  value:
                    answer: "CiteFlowAI pays sources at the moment of citation, in USDC, with unspent budget refunded automatically."
                    citationsUsed:
                      - id: "3f2a91c4-8b7d-4e11-9a03-77c5de10bb42"
                        title: "CiteFlowAI: where every citation pays its author"
                        url: "https://community.arc.io/public/forum/boards/ecosystem-showcase-and-launches-pdq/posts/citeflowai-where-every-citation-pays-its-author-6zuclwmgdh"
                        content: "Every AI research tool has the same unpaid bill…"
                        receipt:
                          payload: "3f2a91c4-8b7d-4e11-9a03-77c5de10bb42:auth_abc123:1785900000000"
                          signature: "9f3c2a1e...  (HMAC-SHA256 hex)"
                          gatewaySettlementId: "de979faf-db16-4a2e-9c31-1a2b3c4d5e6f"
                    purchasedSources:
                      - id: "3f2a91c4-8b7d-4e11-9a03-77c5de10bb42"
                        title: "CiteFlowAI: where every citation pays its author"
                        url: "https://community.arc.io/public/forum/boards/ecosystem-showcase-and-launches-pdq/posts/citeflowai-where-every-citation-pays-its-author-6zuclwmgdh"
                        content: "Every AI research tool has the same unpaid bill…"
                        receipt:
                          payload: "3f2a91c4-8b7d-4e11-9a03-77c5de10bb42:auth_abc123:1785900000000"
                          signature: "9f3c2a1e...  (HMAC-SHA256 hex)"
                          gatewaySettlementId: "de979faf-db16-4a2e-9c31-1a2b3c4d5e6f"
                    transaction: "0x4a1b2c3d..."
                settlement_warning:
                  summary: Answer served, x402 settlement itself failed after the fact
                  value:
                    answer: "…"
                    citationsUsed: []
                    purchasedSources: []
                    settlementWarning: "facilitator unreachable"
        '400':
          description: Missing or too-short query parameter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: 'Missing or too-short "q" query parameter (min 5 characters)'
        '402':
          description: >
            Payment required. Standard x402 PaymentRequired challenge (v1),
            describing the price, network, asset, and recipient needed to
            retry the request with a valid PAYMENT-SIGNATURE header.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/X402PaymentRequired'
              example:
                x402Version: 1
                accepts:
                  - scheme: exact
                    network: "eip155:5042002"
                    maxAmountRequired: "1000000"
                    resource: /api/agent/research
                    description: "CiteFlow AI grounded research answer, agent-payable via x402"
                    mimeType: application/json
                    payTo: "0x933a2405f84c224be1ef373ba16e992e1f459682"
                    maxTimeoutSeconds: 604800
                    asset: "0x...(USDC on Arc Testnet)"
        '500':
          description: Session creation failed, or the research agent execution failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CitedSource:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        url:
          type: string
          format: uri
        content:
          type: string
          description: The retrieved chunk(s) of source content used to ground the answer.
        receipt:
          type: object
          description: Cryptographic proof of the citation payment.
          properties:
            payload:
              type: string
              description: "sourceId:authorizationId:timestamp"
            signature:
              type: string
              description: HMAC-SHA256 signature over payload, hex-encoded.
            gatewaySettlementId:
              type: string
              description: Circle Gateway batch settlement identifier.
          required: [payload, signature, gatewaySettlementId]
      required: [id, title, url, content, receipt]

    ResearchResult:
      type: object
      properties:
        answer:
          type: string
          description: The synthesized answer, grounded exclusively in cited sources.
        citationsUsed:
          type: array
          description: The subset of purchasedSources whose IDs the model actually cited in the answer.
          items:
            $ref: '#/components/schemas/CitedSource'
        purchasedSources:
          type: array
          description: Every source paid for during this session (a superset of citationsUsed in edge cases).
          items:
            $ref: '#/components/schemas/CitedSource'
        transaction:
          type: string
          description: On-chain transaction reference for the settled x402 payment, when settlement succeeded.
        settlementWarning:
          type: string
          description: Present only if the answer was served but x402 settlement failed afterward.
      required: [answer, citationsUsed, purchasedSources]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    X402PaymentRequirement:
      type: object
      description: One accepted payment method, per the x402 v1 spec.
      properties:
        scheme:
          type: string
          example: exact
        network:
          type: string
          example: "eip155:5042002"
        maxAmountRequired:
          type: string
          description: Price in the asset's base units (USDC, 6 decimals). "1000000" = $1.00.
        resource:
          type: string
        description:
          type: string
        mimeType:
          type: string
        payTo:
          type: string
          description: Treasury wallet address receiving payment.
        maxTimeoutSeconds:
          type: integer
        asset:
          type: string
          description: USDC contract address on the given network.
        extra:
          type: object
          nullable: true
          description: Scheme-specific extras (e.g. Circle Gateway's EIP-712 signing domain).
      required: [scheme, network, maxAmountRequired, resource, description, payTo, maxTimeoutSeconds, asset]

    X402PaymentRequired:
      type: object
      properties:
        x402Version:
          type: integer
          enum: [1]
        error:
          type: string
        accepts:
          type: array
          items:
            $ref: '#/components/schemas/X402PaymentRequirement'
      required: [x402Version, accepts]